Deleted circular dep and used context for firebase calls
